pyproject.toml
[tool.poetry]
name = "pyLYNX"
version = "1.2"
description = "A basic EULYNX implementation for Python"
authors = ["Philipp Schmidt <philipp.schmidt@student.hpi.de>"]
readme = "README.md"
license = "GPL-3.0-or-later"
packages = [
    { include = "pyLYNX" },
]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.43"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"